+++ headers/PlayerOne.h
@@ -5,7 +5,7 @@
 class PlayerOne final : public Pawn
 {
 public:
-	PlayerOne(const Point& pos, int width, int height, int color, int speed, int health, Environment* env);
+	PlayerOne(const Point& pos, int width, int height, int color, float speed, int health, Environment* env);
 
 	~PlayerOne() override;
